Fixing compilation issues

+++ src/Utils/ConfigurationUtils.cpp
@@ -230,9 +230,9 @@ vis::ConfigurationUtils::read_color_lines(bool is_override_terminal_colors,
                 vis::Utils::hex_to_int(color_line.substr(1, 6));
 
             // ncurses uses colors between 0-1000, so scale from 0-256 to 0-1000
-            const int16_t red = std::round((static_cast<double>((hex_color >> 16) % 256) / 256.0) * 1000.0);
-            const int16_t green = std::round((static_cast<double>((hex_color >> 8) % 256) / 256.0) * 1000.0);
-            const int16_t blue = std::round((static_cast<double>(hex_color % 256) / 256.0) * 1000.0);
+            const int16_t red = static_cast<int16_t>(std::round((static_cast<double>((hex_color >> 16) % 256) / 256.0) * 1000.0));
+            const int16_t green = static_cast<int16_t>(std::round((static_cast<double>((hex_color >> 8) % 256) / 256.0) * 1000.0));
+            const int16_t blue = static_cast<int16_t>(std::round((static_cast<double>(hex_color % 256) / 256.0) * 1000.0));
 
             // skip color 0, since it is reserved by the terminal for
             // white/black
